concat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -30,7 +30,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	{
 		p[b] = s1[b];
 	}
-	if (n < y)
+	if ((int)n < y)
 	{
 	for (b = i; b <= ((int) n + i); b++)
 	{
